Move api calls to separate file
