+++ App.js
@@ -1,51 +1,38 @@
-import { StyleSheet, Text, View, ImageBackground, Image } from 'react-native'
-import LandingPage from './src/components/LandingPage'
+import { StyleSheet, Text, View, Draggable, StatusBar } from 'react-native'
 import { useState } from 'react'
+import LandingPage from './src/components/LandingPage'
+import { getLinks, getFeatured } from './apiCalls'
 
 function App() {
   const [onPage, setOnPage] = useState('landing')
   const [featuredTitle, setFeaturedTitle] = useState('')
   const [featuredLinks, setFeaturedLinks] = useState('')
 
-async function getFeatured() {
-  const currentDate = new Date();
-  const year = currentDate.getFullYear()
-  const month = String(currentDate.getMonth() + 1).padStart(2, '0')
-  const day = String(currentDate.getDate()).padStart(2, '0')
-  const parsedDate = `${year}/${month}/${day}`
-  const res = await fetch(`https://en.wikipedia.org/api/rest_v1/feed/featured/${parsedDate}`)
-  return res.json()
-}
-
-async function getLinks(title) {
-  const res = await fetch(`https://en.wikipedia.org/api/rest_v1/page/related/${title}`)
-  return res.json()
-}
-
-async function startGame() {
- setOnPage('game')
-  try {
-    const data = await getFeatured()
-    setFeaturedTitle(data.tfa.title)
-    try{
-      const links = await getLinks()
-      setFeaturedLinks(links.links)
+  async function startGame() {
+    setOnPage('game')
+    try {
+      const data = await getFeatured()
+      const title = data.tfa.title
+      setFeaturedTitle(title)
+      try {
+        const links = await getLinks(title)
+        setFeaturedLinks(links)
+      } catch (error) {
+        console.error('An error occurred:', error)
+      }
     } catch (error) {
       console.error('An error occurred:', error)
     }
-  } catch (error) {
-    console.error('An error occurred:', error)
   }
-}
 
   function showHelp() {
     setOnPage('help')
   }
 
-
-
   return (
     
+    <>
+      <StatusBar hidden={true} />
       <View style={styles.mainContainer}>
         {onPage === 'landing' &&
           <LandingPage startGame={startGame} showHelp={showHelp}/>
@@ -53,14 +40,14 @@ async function startGame() {
         {onPage === 'game' &&
           <View style={{width: '100%', textAlign: 'center', flexDirection: 'column'}}>
             <Text>{featuredTitle}</Text>
-            <Text>{featuredLinks}</Text>
+            <View style={styles.linksContainer}>{featuredLinks}</View>
           </View>
         }
         {onPage === 'help' &&
           <View></View>
         }
       </View>
-
+    </>
   )
 }
 
@@ -69,5 +56,8 @@ export default App
 const styles = StyleSheet.create({
   mainContainer: {
     flex: 1
+  },
+  linksContainer: {
+    flexDirection: 'column'
   }
 })
